More formatting changes/code reworks *Removed warning about data loss due to conversion +Renamed WIDTH and HEIGHT to WINDOW_WIDTH and WINDOW_HEIGHT in globalConst.h *Formatting changes *Added a 'p' to member variables that are pointers *Modifed the names of the variables inside the KeyInput struct +Color now holds a pointer to ALLEGRO_COLOR

+++ GameAI/Decision/pathing/pathing.cpp
@@ -7,12 +7,10 @@
 
 int pathing::heurDistance(Node* start, Node* goal)
 {
-	int x = (start->x - goal->x);
-	int y = (start->y - goal->y);
+	float x = (float) (start->x - goal->x);
+	float y = (float) (start->y - goal->y);
 	return (int) sqrtf(x * x - y * y);
 }
-
-#include <iostream>
 
 std::vector<Node> pathing::aStar(Grid* grid, Node* start, Node* goal, Heuristic heur)
 {
